Refactored writers, introduced WriterBase (#1)
* Refactored writer to be based of a generic writer base

WebDatasetWriter inherits from WriterBase.
Specific writers do not have to worry about end of sequence and end of queue tokens, but can override if wanted
Pipeline has an additional method to() for attaching a writer. A pipeline can only have one writer. The writer is automatically put at the end.