Added memory writers for small amounts of data that need to be patched up (i.e. in scripts).
Added both a TorchMemoryWriter and a NumpyMemoryWriter (including demo scripts)

These writers can be used for in memory patching in workflows. 

Also added ContextAware Mixin for writer and stage classes. And added ctx to writer stages (so you can use the use_gpu ctx and others in your writer if necessary).